+++ python/src/pshell.py
@@ -68,6 +68,7 @@ _gHelp = ('?', '-h', '--h', '-help', '--help', 'help')
 
 _gFileSystemPath = "/tmp/"
 _gLockFileExtension = ".pshell-lock"
+_gUnixLockFileId = "unix"+_gLockFileExtension
 _gActiveServers = []
 
 #################################################################################
@@ -406,6 +407,7 @@ def _cleanupFileSystemResources():
   global _gActiveServers
   global _gMaxHostnameLength
   global _gMaxActiveServerLength
+  global _gUnixLockFileId
   lockFiles = fnmatch.filter(os.listdir(_gFileSystemPath), "*"+_gLockFileExtension)
   lockFiles.sort()
   for file in lockFiles:
@@ -416,7 +418,8 @@ def _cleanupFileSystemResources():
         # we got the lock, delete any socket file and lock file and don't print anything
         try:
           # unix temp socket file
-          os.unlink(_gFileSystemPath+file.split("-")[0])
+          if _gUnixLockFileId in file:
+            os.unlink(_gFileSystemPath+file.split("-")[0])
         except:
           None
         try:
